+++ src/main/java/com/samoonpride/backend/repository/MediaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MediaRepository extends CrudRepository<Media, Long> {
+public interface MediaRepository extends CrudRepository<Media, Integer> {
 }
